#607483 - moving calculateImportString into a separate file
